Move tests for problem 1.5 to combined chapter test.
